Correctly handling of function pointer types.
Changed macros to be consistent with how C macros
are supposed to be defined, aka upper case.
Keeping the identifier unmodified in the macro to increased readability
and correctness.

The approach of moving the identifier to the macro increases the
flexibility of initialisation by allowing the user to do "whatever"
with the identifier.