test: cleanup and stabilize TestEnv setup
tryRemove can fail thus need to catch the exception.
Not terribly important when it fails, usually something with the disc I/O.

Only need to create the outdir when it doesn't exist.
Because the tests can run in parallel it is important to only create
the directory when it doesn't exist.