Improve location analyze to allow clear distinguish between a
symbols definition and location.
It is used in the UML diagrams for the creation of relations between
components.

This is the first step in separating the location from the TypeKind.
When the separation is fully completed it allows memory optimization.
The location of symbols is a memory thief number one.