ctestdouble: Fix bug with multiple --in and --file-exclude
See the test case for an example.
Basically the algorithm for header_filtering was constructed for only
running one file at a time.

With the new algorithm it goes through a collection of locations where a
root may only override the current temporary collection.
Between each pass the collected data is moved to a permanent pool.