test: add an extensible dextool executioner
Problem:
The use of the dextool executioners runDextool/runDextool2 has a high
probability of breaking if the parameter list changes.
The pattern of a plain function also has the tendency to _grow_ in the
number of paramters until it become unwieldy.

The new one that this commit add uses the builder pattern to correct
this problem. The builder patterns make it possible to add _new_ ways of
running deXtool without breaking any tests.

By using the builder pattern it is possible to extend without breaking
_older_ usage of the dextool executioner. The older ones with a fixed
parameter list has the unfortunate design limitation that adding more
parameters either break older usage or the parameter list become
unwieldy.