mutate: fix how a path is tested to be inside another path
Paths containing symlinks resolve differently on different systems which
lead to the check of comparing the beginning of strings to see if a path
is inside a directory structure mysteriously fail from the users
perspective.

By converting the paths to the resolved, real, absolute path before
comparing it fixes the bug.

The problem that may now happen is if the user intentionally have
symlinks that the user wants to restrict inside.