mutate: fix cleanup of temp dirs on NFS systems
There can be stale file handlers in a tmp directory which mean that at
that specific time the directory can't be removed. But it will *most
probably* succeed at a later time when the NFS has synced up.

By keeping a list of temp directories and continuously trying to remove
them most cases of junk being left behind should be fixed.