mutate: mutation schema DCR do not mutate predicates
This was hidden because one of the most common predicates is a binary
operator. The binary operator expression is treated as a clause which
results in the same source code mutation.

This commit contains a test case that test for a predicate mutation.

The problem though is that it results in duplicate source code mutations
and thus adversely affects the runtime.