mutate: nomut is not always show in the html report
The reason is that a mutant that span multiple lines may have the NOMUT
comment on a line different from the where it starts. Thus the
optimization that where used which only store the NOMUT per line failed.

The solution is to query the database for each mutant and get the
metadata for it. The con is that this is a bit slower but it works.