mutate: workaround that a schema that contains mutants isn't executed in
the test phase.
This is because the schema tables allow the same mutation status ID to
be added multiple times with a relation to the schema. The InjectBuilder
in turn then removes mutants if they occur multiple times. The result is
that all mutants are removed just before the schema is executed.